+++ lib/museum/publication/publicEntityGraphMedia.ts
@@ -41,6 +41,7 @@ import {
   metadataOnlyMedia,
   proposalMetadata,
 } from "./publicEntityGraphMediaMetadata";
+import { accessionMediaFacts } from "./publicEntityGraphAccessionMedia";
 export { findWavePublicationPart } from "./publicEntityGraphWaveMediaJoin";
 
 export function mapWorkStatus(
@@ -302,17 +303,7 @@ function projectProposalMedia(
   assertProposalPresentationInput(input);
   assertProposalContext(media, subjectEntity, context, waveId, dropId);
   assertProposalSourceLocator(input);
-  const sourceByteSize = requiredPositiveInteger(
-    media,
-    "source_byte_size",
-    "public_entity_graph_media_source_byte_size"
-  );
-  const publicationPartNumber = requiredPositiveInteger(
-    media,
-    "publication_part_number",
-    "public_entity_graph_media_publication_part_number"
-  );
-  assertAccessionMediaRecord(
+  const { sourceByteSize, publicationPartNumber } = accessionMediaFacts(
     mediaEntity,
     media,
     publicationRecordId
@@ -357,69 +348,6 @@ function projectProposalMedia(
     throw new Error("public_entity_graph_media_proposal_contract");
   }
   return { retained: null, presentation: candidate, metadata: null };
-}
-
-function requiredPositiveInteger(
-  source: Readonly<Record<string, unknown>>,
-  key: string,
-  error: string
-): number {
-  const value = source[key];
-  if (
-    typeof value !== "number" ||
-    !Number.isSafeInteger(value) ||
-    value < 1
-  ) {
-    throw new Error(error);
-  }
-  return value;
-}
-
-function assertAccessionMediaRecord(
-  mediaEntity: MuseumPublicEntityRecord,
-  media: Readonly<Record<string, unknown>>,
-  publicationRecordId: string
-): void {
-  const context = requiredObject(
-    media,
-    "wave_proposal_context",
-    "public_entity_graph_media_wave"
-  );
-  const observationRecordId = requiredString(
-    context,
-    "observation_record_id",
-    "public_entity_graph_media_observation"
-  );
-  const fixity = requiredObject(
-    media,
-    "fixity",
-    "public_entity_graph_media_fixity"
-  );
-  const rights = requiredObject(
-    media,
-    "rights",
-    "public_entity_graph_media_rights"
-  );
-  const sourceObservation = requiredObject(
-    media,
-    "source_observation",
-    "public_entity_graph_media_source_observation"
-  );
-  if (
-    context["publication_status"] !==
-      "historical_public_proposal_context" ||
-    !mediaEntity.sourceRecordIds.includes(publicationRecordId) ||
-    !mediaEntity.sourceRecordIds.includes(observationRecordId) ||
-    fixity["status"] !== "verified" ||
-    fixity["algorithm"] !== "sha256" ||
-    typeof fixity["digest"] !== "string" ||
-    !/^sha256:[a-f0-9]{64}$/u.test(fixity["digest"]) ||
-    typeof fixity["verified_at"] !== "string" ||
-    rights["status"] !== "restricted" ||
-    sourceObservation["status"] !== "mutable_external"
-  ) {
-    throw new Error("public_entity_graph_media_accession_record");
-  }
 }
 
 function assertProposalPresentationInput(
